Refactor noteDetail.css and note.html
This commit refactors the noteDetail.css and note.html files. It includes changes to the CSS styles in noteDetail.css, such as removing the background color of the btn-primary class and adjusting the styles of the section-note class. In note.html, a new paragraph with the content label is added.